// flogui/src/app/client-spacing/models/product.interface.ts
export interface Products {
    id: any;
    type: string;
    forme: string;
    couleur: string;
    couleurdeux: string;
    couleurtrois: string;
    materiaux: string;
    materiauxdeux: string;
    materiauxtrois: string;
    prix: number;
    image: any;
    information: string;
    stock: number;
}